requirements.txt
aiofiles==22.1.0
aiogram==3.0.0b6
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.3
psycopg2==2.9.5
pydantic==1.10.2
python-dotenv==0.21.0
pytz==2022.6
typing_extensions==4.4.0
yarl==1.8.2
